Add all pre- and post-process related libraries and test files
