+++ test/test_pre_processes.rb
@@ -1,3 +1,6 @@
+# frozen_string_literal: true
+
+require_relative '../lib/pre.d/combine_transpile_js.rb'
 
 class TestJsCombineAndTranspile < Test::Unit::TestCase
   ##
@@ -34,7 +37,7 @@ class TestJsCombineAndTranspile < Test::Unit::TestCase
     html_footer_filepath = write_tempfile('docinfo_footer.html', html_footer)
 
     docinfo_files_paths = OpenStruct.new('header' => html_header_filepath, 'footer' => html_footer_filepath)
-    results = Toolchain::CombineAndTranspileJs.new.run(docinfo_files_paths)
+    results = Toolchain::Pre::CombineAndTranspileJS.new.run(docinfo_files_paths)
     assert_equal(Zlib.crc32(results[0].js_blob), 2284332409)
     assert_equal(Zlib.crc32(results[0].html), 3050793018)
     assert_equal(Zlib.crc32(results[1].js_blob), 1211914232)
